refactor: apply the over-engineering review of the interact-key migration
No behaviour change except where noted.

- Move isTypingTarget out of MicrophoneSessionStore into Utils/CustomTypeGuards:
  a keyboard guard had no business living in a microphone store, and that
  misplacement is exactly why the two pre-existing copies of the same check could
  never reuse it. Its test moves with it.
- Collapse those two copies onto the helper (UserInputManager's per-tick key poll
  and the Ctrl+D "walk my desk" guard). They also gain contenteditable coverage,
  which the raw triple-instanceof did not have — the one behaviour change here.
- Drop the dead Math.max(24, ...) clamp on the keycap width: INTERACT_KEY is at
  least one character, so the expression is always >= 26 and the branch is
  unreachable. Drop the single-use `label` alias with it.
- Rename handlePressInteractKeyPlayerTextCallback to firePlayerTextCallbacks:
  the old name described the key that triggers it, so it needed renaming every
  time the shortcut moves. It fires every playText callback on the player.
- Halve the two comment blocks in handleKeyUpEvent that were longer than the two
  lines of code they explained.

Kept deliberately: capWidth stays derived from INTERACT_KEY rather than hardcoded
to 26. The constant just moved once, and a hardcoded cap would silently mis-size
the badge the next time it does.
