+++ play/src/front/Phaser/Entity/Character.ts
@@ -773,7 +773,7 @@ export abstract class Character extends Container implements OutlineableInterfac
         return this._lastDirection;
     }
 
-    public handlePressInteractKeyPlayerTextCallback() {
+    public firePlayerTextCallbacks() {
         for (const [, text] of this.texts) {
             (text as SpeechDomElement).callback();
         }
