+++ play/src/front/Utils/CustomTypeGuards.ts
@@ -11,3 +11,18 @@ export function isOutlineable(object: unknown): object is OutlineableInterface {
 export function assertObjectKeys<T extends object>(obj: T): (keyof T)[] {
     return Object.keys(obj) as Array<keyof T>;
 }
+
+/**
+ * Guard so a game shortcut does not fire while the user is typing: the key must keep its normal
+ * text-input behaviour in text fields and contenteditable elements.
+ */
+export function isTypingTarget(target: EventTarget | null): boolean {
+    return (
+        target instanceof HTMLInputElement ||
+        target instanceof HTMLTextAreaElement ||
+        target instanceof HTMLSelectElement ||
+        // isContentEditable covers nested nodes; the attribute check is the jsdom fallback (jsdom does
+        // not implement isContentEditable).
+        (target instanceof HTMLElement && (target.isContentEditable || target.contentEditable === "true"))
+    );
+}
